requirements.txt
apache-airflow
requests
beautifulsoup4
sentence-transformers
pymilvus
pyspark
